refactore: partioning aggregator functions
