+++ package.json
@@ -6,7 +6,8 @@
     "fetch": "tsx fetch",
     "report": "tsx report",
     "lint:type": "tsc --noEmit",
-    "lint:eslint": "eslint **/*.ts"
+    "lint:eslint": "eslint **/*.ts",
+    "test": "vitest"
   },
   "dependencies": {
     "@gitbeaker/node": "^35.6.0",
@@ -16,7 +17,8 @@
     "glob": "^8.0.3",
     "glob-promise": "^4.2.2",
     "got": "^12.1.0",
-    "sqlite3": "^5.0.8"
+    "sqlite3": "^5.0.8",
+    "tsx": "^3.7.0"
   },
   "devDependencies": {
     "@types/glob": "^7.2.0",
@@ -25,7 +27,6 @@
     "eslint": "^8.19.0",
     "eslint-config-prettier": "^8.5.0",
     "prettier": "^2.7.1",
-    "tsx": "^3.7.0",
     "typescript": "^4.7.4",
     "vitest": "^0.16.0"
   }
